+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Support for `else if` chains in `cfg_iif!` macro.
+* Support for multiple `cfg` predicates in `cfg_iif!` macro (implicitly combined with `all()`).
+* More examples and tests for the new features.
+
+### Fixed
+* Bug where multiple `cfg` predicates caused a compilation error in `else` branches.
 
 
 ## [0.2.8] (2025-09-24)